feat: VideoFunction분리
