--- src/pages/Search/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Search = () => {
-  return <div>Search</div>;
-};
-
-export default Search;
