+++ src/router.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from 'react-router-dom';
 import App from './App';
-import Search from './pages/Search';
-import Home from './pages/Home';
+import Search from './pages/Search/Search';
+import Home from './pages/Home/Home';
 import Detail from './pages/Detail';
 
 const router = createBrowserRouter([
